# contracts/review.openapi.yaml
openapi: 3.1.0
info:
  title: RFA review contract (결재 문서)
  version: 0.1.0
  description: |
    결재 문서의 단일 진실 원천. 상태기계로 순서를 강제한다.
      opened → knowledge_ready → drafted → scanned → reviewed → approved → posted
                                                              └→ rejected
      (어느 단계든) → needs_human
    순서 위반은 409. approve/reject 는 호스트 loopback(사람)만 호출할 수 있다.
    구현 단계: Step 3(코어), Step 4(scanned, policy), Step 5(approve/reject).
servers:
  - url: http://127.0.0.1:8790
paths:
  /reviews:
    post:
      operationId: openReview
      summary: 결재 문서 개설 (→ opened)
      parameters: [{ $ref: "#/components/parameters/Actor" }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/OpenReviewRequest" }
      responses:
        "201":
          description: 생성된 문서
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Review" }
    get:
      operationId: listReviews
      parameters:
        - name: status
          in: query
          schema: { $ref: "#/components/schemas/ReviewStatus" }
      responses:
        "200":
          description: 목록
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Review" }
  /reviews/{id}:
    get:
      operationId: getReview
      parameters: [{ $ref: "#/components/parameters/ReviewId" }]
      responses:
        "200":
          description: 문서 전체
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Review" }
        "404": { $ref: "#/components/responses/NotFound" }
  /reviews/{id}/knowledge:
    post:
      operationId: attachKnowledge
      summary: 실무대장 응답 첨부 (opened → knowledge_ready)
      parameters: [{ $ref: "#/components/parameters/ReviewId" }, { $ref: "#/components/parameters/Actor" }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "knowledge.openapi.yaml#/components/schemas/KnowledgeResult" }
      responses:
        "200": { $ref: "#/components/responses/ReviewOk" }
        "409": { $ref: "#/components/responses/InvalidTransition" }
  /reviews/{id}/draft:
    post:
      operationId: submitDraft
      summary: 최종 초안 제출. 서버가 비밀값 스캔을 자동 실행 (knowledge_ready → drafted → scanned)
      parameters: [{ $ref: "#/components/parameters/ReviewId" }, { $ref: "#/components/parameters/Actor" }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [text]
              properties:
                text: { type: string }
                edit_log:
                  type: array
                  items: { $ref: "#/components/schemas/EditVerdict" }
      responses:
        "200": { $ref: "#/components/responses/ReviewOk" }
        "409": { $ref: "#/components/responses/InvalidTransition" }
  /reviews/{id}/verdict:
    post:
      operationId: submitVerdict
      summary: 기밀검토 판정 (scanned → reviewed)
      parameters: [{ $ref: "#/components/parameters/ReviewId" }, { $ref: "#/components/parameters/Actor" }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/Verdict" }
      responses:
        "200": { $ref: "#/components/responses/ReviewOk" }
        "409": { $ref: "#/components/responses/InvalidTransition" }
  /reviews/{id}/needs-human:
    post:
      operationId: markNeedsHuman
      summary: 자동 처리 불가. 사람에게 넘김 (→ needs_human)
      parameters: [{ $ref: "#/components/parameters/ReviewId" }, { $ref: "#/components/parameters/Actor" }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [reason]
              properties:
                reason: { type: string }
      responses:
        "200": { $ref: "#/components/responses/ReviewOk" }
        "409": { $ref: "#/components/responses/InvalidTransition" }
  /reviews/{id}/approve:
    post:
      operationId: approveReview
      summary: 사람 승인 → 서명 토큰 발급 → 게시 (reviewed → approved → posted). loopback 전용
      parameters: [{ $ref: "#/components/parameters/ReviewId" }]
      responses:
        "200":
          description: 게시 결과
          content:
            application/json:
              schema:
                type: object
                required: [status]
                properties:
                  status: { $ref: "#/components/schemas/ReviewStatus" }
                  posted_url: { type: string, format: uri }
        "403": { description: "loopback 이 아닌 출처. {error: loopback_only}" }
        "409": { description: "순서 위반(invalid_transition) / verdict=block(verdict_is_block) / 토큰 잔존(secrets_in_final_body)" }
        "502": { description: "게시 실패. {error: publish_failed, detail}. 문서는 approved 로 남는다" }
  /reviews/{id}/reject:
    post:
      operationId: rejectReview
      summary: 사람 거절. 사유는 feedback 으로 축적 (reviewed → rejected). loopback 전용
      parameters: [{ $ref: "#/components/parameters/ReviewId" }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [reason]
              properties:
                reason: { type: string }
      responses:
        "200": { $ref: "#/components/responses/ReviewOk" }
        "403": { description: loopback 이 아닌 출처 }
        "409": { $ref: "#/components/responses/InvalidTransition" }
  /policy/{scope}:
    get:
      operationId: getPolicy
      summary: 기밀 기준 (official + personal + 최근 feedback)
      parameters:
        - name: scope
          in: path
          required: true
          schema: { type: string, enum: [public, internal] }
      responses:
        "200":
          description: 기준
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Policy" }
        "404":
          description: 없는 scope 이거나 그 scope 의 official.md 가 아직 없음
          content:
            application/json:
              schema:
                type: object
                required: [error, scope]
                properties:
                  error: { type: string, const: not_found }
                  scope: { type: string }
components:
  parameters:
    Actor:
      name: X-RFA-Actor
      in: header
      required: false
      description: 이 요청을 보낸 주체(intake, press, censor_public 등). events[].who 에 기록. 없으면 "unknown"
      schema: { type: string, default: unknown }
    ReviewId:
      name: id
      in: path
      required: true
      schema: { type: integer, minimum: 1 }
  responses:
    NotFound:
      description: 없는 결재 문서
      content:
        application/json:
          schema:
            type: object
            required: [error, id]
            properties:
              error: { type: string, const: not_found }
              id: { type: integer }
    ReviewOk:
      description: 갱신된 문서
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Review" }
    InvalidTransition:
      description: 순서 위반
      content:
        application/json:
          schema:
            type: object
            required: [error, from, to]
            properties:
              error: { type: string, const: invalid_transition }
              from: { $ref: "#/components/schemas/ReviewStatus" }
              to: { $ref: "#/components/schemas/ReviewStatus" }
  schemas:
    Channel: { type: string, enum: [public, internal] }
    ReviewStatus:
      type: string
      enum: [opened, knowledge_ready, drafted, scanned, reviewed, approved, rejected, posted, needs_human]
    OpenReviewRequest:
      type: object
      required: [channel, target, source_url, requester, question]
      properties:
        channel: { $ref: "#/components/schemas/Channel" }
        target: { type: string, pattern: "^[A-Za-z0-9_.-]+/[A-Za-z0-9_.-]+#[1-9][0-9]*$", examples: ["zetwhite/rfa-test#34"] }
        source_url: { type: string, format: uri }
        requester: { type: string }
        question: { type: string }
    EditVerdict:
      type: object
      required: [round, verdict]
      properties:
        round: { type: integer, minimum: 1 }
        verdict: { type: string, enum: [pass, revise] }
        notes: { type: string }
        issues: { type: array, items: { type: string } }
    ScanHit:
      type: object
      required: [type, match, span]
      properties:
        type: { type: string, enum: [token, private_ip, internal_host, internal_path] }
        match: { type: string }
        span: { type: array, items: { type: integer }, minItems: 2, maxItems: 2 }
    VerdictReason:
      type: object
      required: [rule, span, action]
      properties:
        rule: { type: string, pattern: "^(official|personal|scanner|editor):[a-z0-9-]+$", examples: ["official:release-date"] }
        span: { type: string, description: 원문 구절 }
        action: { type: string, enum: [remove, blur, keep] }
    Verdict:
      type: object
      required: [verdict, summary]
      properties:
        verdict: { type: string, enum: [allow, redact, block] }
        redacted_body: { type: string, description: redact 일 때 필수 }
        reasons: { type: array, items: { $ref: "#/components/schemas/VerdictReason" } }
        summary: { type: string }
    Event:
      type: object
      required: [at, who, what]
      properties:
        at: { type: string, format: date-time }
        who: { type: string }
        what: { type: string }
        detail: { type: string }
    Decision:
      type: object
      required: [by, at]
      properties:
        by: { type: string }
        at: { type: string, format: date-time }
        reason: { type: string }
    Review:
      type: object
      required: [id, status, channel, target, source_url, requester, question, edit_log, scan, events]
      properties:
        id: { type: integer }
        status: { $ref: "#/components/schemas/ReviewStatus" }
        channel: { $ref: "#/components/schemas/Channel" }
        target: { type: string }
        source_url: { type: string, format: uri }
        requester: { type: string }
        question: { type: string }
        knowledge: { $ref: "knowledge.openapi.yaml#/components/schemas/KnowledgeResult" }
        draft: { type: string }
        edit_log: { type: array, items: { $ref: "#/components/schemas/EditVerdict" } }
        scan: { type: array, items: { $ref: "#/components/schemas/ScanHit" } }
        verdict: { $ref: "#/components/schemas/Verdict" }
        final_body: { type: string, description: redact 면 redacted_body, allow 면 draft }
        decision: { $ref: "#/components/schemas/Decision" }
        events: { type: array, items: { $ref: "#/components/schemas/Event" } }
    FeedbackItem:
      type: object
      required: [at, scope, review_id, decision]
      properties:
        at: { type: string, format: date-time }
        scope: { $ref: "#/components/schemas/Channel" }
        review_id: { type: integer }
        decision: { type: string, enum: [approve, reject] }
        reason: { type: string }
        draft_excerpt: { type: string }
        rule_hint: { type: string }
    Policy:
      type: object
      required: [scope, official, personal, feedback]
      properties:
        scope: { $ref: "#/components/schemas/Channel" }
        official: { type: string, description: official.md 본문 }
        personal: { type: string, description: personal.md 본문 }
        feedback:
          type: array
          description: 같은 scope 의 사람 결재 사례. reject 먼저, 그 안에서 최신순, 최대 10건
          items: { $ref: "#/components/schemas/FeedbackItem" }